requirements.txt
discord==1.7.3
discord.ext==0.8
discord.py==1.7.3
python-dotenv==0.20.0
requests==2.27.1